+++ core/amber/src/main/python/core/architecture/sendsemantics/broad_cast_partitioner.py
@@ -20,7 +20,7 @@ from typing import Iterator
 from overrides import overrides
 from core.architecture.sendsemantics.partitioner import Partitioner
 from core.models import Tuple
-from core.models.marker import State
+from core.models.state import State
 from core.util import set_one_of
 from proto.edu.uci.ics.amber.engine.architecture.sendsemantics import (
     Partitioning,
