+++ core/amber/src/main/python/core/models/payload.py
@@ -17,7 +17,7 @@
 
 from dataclasses import dataclass
 from pyarrow.lib import Table
-from core.models.marker import State
+from core.models.state import State
 
 
 @dataclass
