+++ core/amber/src/main/python/core/runnables/main_loop.py
@@ -38,7 +38,7 @@ from core.models.internal_queue import (
     ChannelMarkerElement,
     InternalQueueElement,
 )
-from core.models.marker import State
+from core.models.state import State
 from core.runnables.data_processor import DataProcessor
 from core.util import StoppableQueueBlockingRunnable, get_one_of
 from core.util.console_message.timestamp import current_time_in_local_timezone
